refactor(editor): controller styled shell 분리
